update PR re. Frank suggested -custom hook
