+++ prettier.config
@@ -1,0 +1,1 @@
+echo {}> .prettierrc.json
